Fix imports after renaming attack-prefix-ids.js to model-names.js
