+++ app/models/tactic-model.js
@@ -2,7 +2,7 @@
 
 const mongoose = require('mongoose');
 const AttackObject = require('./attack-object-model');
-const { ModelName } = require('../lib/attack-prefix-ids');
+const { ModelName } = require('../lib/model-names');
 
 const stixTactic = {
     // STIX x-mitre-tactic specific properties
